acrescentando algumas modificações visuais

+++ autoclean-linux.sh
@@ -10,7 +10,7 @@
 # http://www.flaviodeoliveira.com.br
 # http://www.youtube.com/flaviodicas
 # http://github.com/oliveiradeflavio
-# oliveiradeflavio@gmail.com
+
 
 if [[ `id -u` -ne 0 ]]; then
 	echo
@@ -23,6 +23,15 @@ if [[ `id -u` -ne 0 ]]; then
     fi
     exit 0
 fi
+
+#Criei variaveis para as cores. Isso facilita a visualização do código depois
+#A variavel "normal" ela serve para voltar a cor padrão do shell.
+verde="\033[1;32m"
+azul="\033[1;34m"
+vermelho="\033[1;31m"
+branco="\033[0;37m"
+NORMAL="\033[m"
+
 
 #verifica qual gerenciador de pacotes Linux a distro utiliza
 packagemanager()
@@ -41,44 +50,48 @@ echo
 cleaning_rpm()
 {
 clear
-if which -a prelink; then
+if which -a prelink && which -a pv; then
 	clear
-	echo "Limpando Cache de Bibliotecas"
+	echo -e "${verde}Limpando Cache de Bibliotecas${NORMAL}"
 	dnf -y clean all
 	dnf -y clean headers
 	dnf -y clean packages
 	echo "--------------------------------------------"
-	echo "Removendo Arquivos (.bak, ~, .tmp) da pasta Home"
+	echo -e "${verde}Removendo Arquivos (.bak, ~, .tmp) da pasta Home${NORMAL}"
 	for i in *~ *.bak *.tmp; do
 		find $HOME -iname "$i" -exec rm -f {} \;
 	done
 	echo "--------------------------------------------"
-	echo "Otimizando as Bibliotecas dos Programas"
-	/etc/cron.daily/prelink
+	echo -e "${verde}Otimizando as Bibliotecas dos Programas${NORMAL}"
+	/etc/cron.daily/prelink | pv -p -e
 	echo "--------------------------------------------"
 	clear
-	echo "Limpeza Concluída"
+	echo -e "${azul}#################################${NORMAL}"
+	echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+	echo -e "${azul}#${vermelho} LIMPEZA CONCLUIDA !!!!	${NORMAL}${azul}#${NORMAL}"
+	echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+	echo -e "${azul}#################################${NORMAL}"
 	sleep 2
 else
 	clear
-	echo -e "Você precisa instalar um(1) programa\n para continuar com a Limpeza."
-	read -p "Deseja instalar o Prelink ? s/n: " -n1 escolha
+	echo -e "Você precisa instalar dois programa\n para continuar com a Limpeza."
+	read -p "Deseja instalar o Prelink e  PV? s/n: " -n1 escolha
 	case $escolha in
 		s|S) echo
-			dnf install prelink -y
+			dnf install prelink pv -y
 			sed -i 's/unknown/yes/g' /etc/default/prelink
 			cleaning_rpm
 			;;
-		n|N) echo
-			echo Saindo, não executando a limpeza...
-			sleep 1
-			exit
-			;;
-		*) echo
-			echo Alternativas incorretas ... Saindo!!!
-			sleep 1
-			exit
-			;;
+			n|N) echo
+				echo -e ${azul}Saindo, não executando a limpeza...${NORMAL}
+				sleep 1
+				exit
+				;;
+			*) echo
+				echo -e ${vermelho}Alternativas incorretas... Saindo!!!${NORMAL}
+				sleep 1
+				exit
+				;;
 	esac
 
 fi
@@ -88,57 +101,61 @@ fi
 cleaning_apt()
 {
 clear
-if which -a prelink && which -a deborphan; then
+if which -a prelink && which -a deborphan && which -a pv; then
 	clear
-	echo "Esvaziando a Lixeira"
+	echo -e "${verde}Esvaziando a Lixeira${NORMAL}"
 	rm -rf /home/$SUDO_USER/.local/share/Trash/files/*
 	echo "--------------------------------------------"
-	echo "Esvaziando os Arquivos Temporários (pasta tmp)"
+	echo -e "${verde}Esvaziando os Arquivos Temporários (pasta tmp)${NORMAL}"
 	rm -rf /var/tmp/*
 	echo "--------------------------------------------"
-	echo "Excluindo Arquivos Inúteis do Cache do Gerenciador de Pacotes (apt)"
+	echo -e "${verde}Excluindo Arquivos Inúteis do Cache do Gerenciador de Pacotes (apt)${NORMAL}"
 	apt-get clean -y
 	echo "--------------------------------------------"
-	echo "Excluindo Pacotes Velhos que não tem utilidade para o Sistema"
+	echo -e "${verde}Excluindo Pacotes Velhos que não tem utilidade para o Sistema${NORMAL}"
 	apt-get autoremove -y
 	echo "--------------------------------------------"
-	echo "Excluindo Pacotes Duplicados"
+	echo -e "${verde}Excluindo Pacotes Duplicados${NORMAL}"
 	apt-get autoclean -y
 	echo "--------------------------------------------"
-	echo "Reparando Pacotes com Problemas"
+	echo -e "${verde}Reparando Pacotes com Problemas${NORMAL}"
 	dpkg --configure -a
 	echo "--------------------------------------------"
-	echo "Removendo Pacotes Órfãos"
+	echo -e "${verde}Removendo Pacotes Órfãos${NORMAL}"
 	apt-get remove $(deborphan) -y ; apt-get autoremove -y
 	echo "--------------------------------------------"
-	echo "Removendo Arquivos (.bak, ~, .tmp) da pasta Home"
+	echo -e "${verde}Removendo Arquivos (.bak, ~, .tmp) da pasta Home${NORMAL}"
 	for i in *~ *.bak *.tmp; do
 		find $HOME -iname "$i" -exec rm -f {} \;
 	done
 	echo "--------------------------------------------"
-	echo "Otimizando as Bibliotecas dos Programas"
-	/etc/cron.daily/prelink
+	echo -e "${verde}Otimizando as Bibliotecas dos Programas${NORMAL}"
+	/etc/cron.daily/prelink | pv -p -e
 	echo "--------------------------------------------"
 	clear
-	echo "Limpeza Concluída ... "
+	echo -e "${azul}#################################${NORMAL}"
+	echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+	echo -e "${azul}#${vermelho} LIMPEZA CONCLUIDA !!!!	${NORMAL}${azul}#${NORMAL}"
+	echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+	echo -e "${azul}#################################${NORMAL}"
 	sleep 2
 else
 	clear
-	echo -e "Você precisa instalar dois programas\n para continuar com a Limpeza."
-	read -p "Deseja instalar o Prelink e o Deborphan? s/n: " -n1 escolha
+	echo -e "${vermelho}Você precisa instalar três programas\n para continuar com a Limpeza.${NORMAL}"
+	read -p "Deseja instalar o Prelink, Deborphan e o PV? s/n: " -n1 escolha
 	case $escolha in
 		s|S) echo
-			apt-get install prelink deborphan -y
+			apt-get install prelink deborphan pv -y
 			sed -i 's/unknown/yes/g' /etc/default/prelink
 			cleaning_apt
 			;;
 		n|N) echo
-			echo Saindo, não executando a limpeza...
+			echo -e ${azul}Saindo, não executando a limpeza...${NORMAL}
 			sleep 1
 			exit
 			;;
 		*) echo
-			echo Alternativas incorretas ... Saindo!!!
+			echo -e ${vermelho}Alternativas incorretas... Saindo!!!${NORMAL}
 			sleep 1
 			exit
 			;;
@@ -147,9 +164,11 @@ else
 fi
 }
 clear
-echo "#################################"
-echo "# Bem vindo ao Autoclean Linux 	#"
-echo "#################################"
+echo -e "${azul}#################################${NORMAL}"
+echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+echo -e "${azul}#${vermelho} BEM VINDO AO AUTOCLEAN LINUX	${NORMAL}${azul}#${NORMAL}"
+echo -e "${azul}#${NORMAL}				${azul}#${NORMAL}"
+echo -e "${azul}#################################${NORMAL}"
 echo
 read -n1 -p "Para continuar escolha s(sim) ou n(não)  " escolha
 	case $escolha in
@@ -157,12 +176,12 @@ read -n1 -p "Para continuar escolha s(sim) ou n(não)  " escolha
 			packagemanager
 			;;
 		n|N) echo
-			echo Finalizando o script...
+			echo -e ${azul}Finalizando o script...${NORMAL}
 			sleep 0.5
 			exit
 			;;
 		*) echo
-			echo Alternativa incorreta!! Saindo...
+			echo -e ${vermelho}Alternativa incorreta!! Saindo...${NORMAL}
 			sleep 1
 			exit
 			;;
